Create report generator integration test
