+++ api/src/test/resources/edu/cornell/mannlib/vitro/webapp/dynapi/components/endpoint_procedure_create_report_generator_input.n3
@@ -1,0 +1,27 @@
+@prefix rdf:  <http://www.w3.org/1999/02/22-rdf-syntax-ns#> .
+@prefix rdfs:  <http://www.w3.org/2000/01/rdf-schema#> .
+@prefix owl:   <http://www.w3.org/2002/07/owl#> .
+@prefix vitro: <http://vitro.mannlib.cornell.edu/ns/vitro/0.7#> .
+@prefix dynapi: <https://vivoweb.org/ontology/vitro-dynamic-api#> .
+@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
+
+<https://vivoweb.org/procedure/create_report/parameter/report_generator_description>
+        dynapi:isInternal                   "true"^^xsd:boolean ;
+        dynapi:defaultValue                 "Dummy report generator description"  .
+        
+<https://vivoweb.org/procedure/create_report/parameter/report_generator_name>
+        dynapi:isInternal                   "true"^^xsd:boolean ;
+        dynapi:defaultValue                 "Dummy report generator name"  .
+
+<https://vivoweb.org/procedure/create_report/parameter/container/construct_queries>
+        dynapi:isInternal                   "true"^^xsd:boolean ;
+        dynapi:defaultValue                 """[{"name": "constructQuery1", "description": "dummy description 1", "constructQuery": "CONSTRUCT Query text 1" }, {"name": "constructQuery2", "description": "dummy description 2", "constructQuery": "CONSTRUCT Query text 2" } ]""" .
+        
+<https://vivoweb.org/procedure/create_report/parameter/container/select_queries>
+        dynapi:isInternal                   "true"^^xsd:boolean ;
+        dynapi:defaultValue                 """[{"name": "data source 1 (select query)", "description": "dummy description select query 1", "selectQuery": "Select query text 1", "graph": "constructQuery1" }, {"name": "data source 2 (select query)", "description": "dummy description 2", "selectQuery": "Select query text 2" , "graph": "constructQuery2"} ]""" .
+        
+
+<https://vivoweb.org/procedure/create_report/parameter/template_base64>
+        dynapi:isInternal                   "true"^^xsd:boolean ;
+        dynapi:defaultValue                 "TEMPLATE_BASE64_ENCODED_STRING_SHOULD_BE_HERE" .
